refactoring to use multi stage docker file for workers
