+++ docker/worker.dockerfile
@@ -1,30 +1,18 @@
-FROM python:3.10
+FROM python:3.10 as common
 
-# Compute the hash of your source files
-ARG SOURCE_HASH=$(find  /config.ini /src/nifty_common /src/nifty_worker -type f -exec sha256sum {} \; | sha256sum | cut -d' ' -f1)
+COPY /src/nifty_worker/requirements.txt /app/requirements.txt
+WORKDIR /app
+RUN python -m pip install -r requirements.txt
 
-# Store the hash in an environment variable
-ENV SOURCE_HASH=$SOURCE_HASH
-
-# Only copy the sources if the hash has changed
 COPY config.ini /app/config.ini
 COPY /src/nifty_common/ /app/nifty_common
-COPY /src/nifty_worker/ /app/nifty_worker
+COPY /src/nifty_worker/__init__.py /app/nifty_worker/__init__.py
+COPY /src/nifty_worker/common/ /app/nifty_worker/common
 
-ONBUILD RUN set -e; \
-    NEW_HASH=$(find /config.ini /src/nifty_common /src/nifty_worker -type f -exec sha256sum {} \; | sha256sum | cut -d' ' -f1); \
-    if [ "$SOURCE_HASH" != "$NEW_HASH" ]; then \
-        # Perform the COPY step for your sources
-        rm -rf /app; \
-        COPY config.ini /app/config.ini; \
-        COPY /src/nifty_common/ /app/nifty_common; \
-        COPY /src/nifty_worker/ /app/nifty_worker; \
-    fi
+FROM common AS worker-trend
 
-# Set the working directory to the app directory
+COPY /src/nifty_worker/trend_worker/ /app/nifty_worker/trend_worker
+
 WORKDIR /app
 
-RUN python -m pip install -r nifty_worker/requirements.txt
-
-# Run the app using Gunicorn
 CMD ["python", "-m", "nifty_worker.trend_worker"]
